Add sitemap generation and cloudinary package
In this commit, the `next-sitemap` package was added in `package.json` to generate a sitemap for the site. Furthermore, the `cloudinary` package also installed to handle image storage and manipulation. A new `postbuild` script added to generate the sitemap after build. A new file `next-sitemap.config.js` was also created to specify the sitemap generation configurations.